+++ mobile/android/android-components/components/feature/downloads/src/main/java/mozilla/components/feature/downloads/DownloadsFeature.kt
@@ -54,8 +54,7 @@ class DownloadsFeature(
      * to be processed.
      */
     override fun start() {
-        val session = sessionId?.let { sessionManager.findSessionById(sessionId) }
-        session?.let { observeFixed(it) } ?: observeSelected()
+        observeIdOrSelected(sessionId)
 
         findPreviousDialogFragment()?.let {
             reAttachOnStartDownloadListener(it)